consolidate identity methods into one class
