+++ app/views/accounts/_connected_app.html.erb
@@ -8,8 +8,9 @@
           <%= identity.display_name %>
         <% end %>
       </div>
+
       <%= t('account.connected_apps.associated',
-            timestamp: local_time(identity.created_at, t('time.formats.event_timestamp'))).html_safe %>
+            timestamp: local_time(identity.created_at.utc, t('time.formats.event_timestamp'))).html_safe %>
     </div>
     <% if identity.service_provider_id %>
       <div class="sm-col sm-col-6 padding-x-1 sm-right-align line-height-4">
