+++ app/views/accounts/connected_accounts/show.html.erb
@@ -11,7 +11,7 @@
 <div class="margin-bottom-4 card profile-info-box border-bottom border-light-blue">
   <div class="border-bottom border-light-blue">
     <% @view_model.connected_apps.each do |identity| %>
-      <%= render identity.connected_app_partial, identity: identity %>
+      <%= render 'accounts/connected_app', identity: identity %>
     <% end %>
   </div>
 </div>
